[BUGFIX] Certificates only generated on first run
The initial certificate generation was only performed if the first in
the list of certificates did not exist. As the renewal is only performed
after a certain amount of days (which we should raise now), we can
safely run the script each time, to ensure we always get new
certificates immediately.